refactor(tangle-dapp): Move path & URL related type defs to their own file
