+++ apps/tangle-dapp/app/claim/success/page.tsx
@@ -1,7 +1,7 @@
 import { isHex } from '@polkadot/util';
 import { redirect } from 'next/navigation';
 
-import { PagePath } from '../../../types';
+import { PagePath } from '../../../types/paths';
 import { getPolkadotApiPromise } from '../../../utils/polkadot';
 import SuccessClient from './SuccessClient';
