+++ apps/tangle-dapp/containers/DelegationsPayoutsContainer/DelegationsPayoutsContainer.tsx
@@ -18,8 +18,9 @@ import { ContainerSkeleton, TableStatus } from '../../components';
 import useDelegations from '../../data/DelegationsPayouts/useDelegations';
 import usePayouts from '../../data/DelegationsPayouts/usePayouts';
 import useIsFirstTimeNominatorSubscription from '../../hooks/useIsFirstTimeNominatorSubscription';
-import useQueryParamKey, { QueryParamKey } from '../../hooks/useQueryParamKey';
+import useQueryParamKey from '../../hooks/useQueryParamKey';
 import { Payout } from '../../types';
+import { QueryParamKey } from '../../types/paths';
 import { convertToSubstrateAddress } from '../../utils';
 import { DelegateTxContainer } from '../DelegateTxContainer';
 import { PayoutAllTxContainer } from '../PayoutAllTxContainer';
